+++ contracts/dragonNFT/interface/IDragonNFT.sol
@@ -9,9 +9,9 @@ interface IDragonNFT {
     function addExperience(uint256 tokenId) external;
     function setXpToLevelUp(uint8 level, uint128 xpRequired) external;
     function getDragonInfo(uint256 tokenId) external view returns(DragonNFTLib.Dragon memory dragonInfo);
-    function getRarityBasedExperience() external view returns (uint32[] memory);
+    function getRarityBasedExperience() external view returns (uint8[] memory);
     function getSpeciesCountPerRarity() external view returns (uint8[] memory);
-    function getRarityBasedDamage() external view returns (uint64[] memory);
+    function getRarityBasedDamage() external view returns (uint16[] memory);
     function getOwnedTokens(address user) external view returns (uint256[] memory);
     function ownerOf(uint256 tokenId) external view returns (address owner);
 }
